Enable Interrupt
In the "enterCritical()" function, the 'cli' command is used to deactivate interrupts. However, in "exitCritical()" it is not reactivated using the 'sei' command. This creates problem when trying to send command to the SMS (use the Supermodified object) from an ISR. The is why the 'sei' command has been added.